+++ Sources/SuperwallKit/Paywall/View Controller/Survey/SurveyManager.swift
@@ -92,7 +92,6 @@ final class SurveyManager {
           paywallInfo: paywallInfo,
           factory: factory,
           paywallViewController: presenter,
-          isDebuggerLaunched: isDebuggerLaunched,
           alertController: alertController,
           completion: completion
         )
@@ -133,7 +132,6 @@ final class SurveyManager {
             paywallInfo: paywallInfo,
             factory: factory,
             paywallViewController: presenter,
-            isDebuggerLaunched: isDebuggerLaunched,
             alertController: otherAlertController,
             completion: completion
           )
@@ -170,37 +168,32 @@ final class SurveyManager {
     paywallInfo: PaywallInfo,
     factory: TriggerFactory,
     paywallViewController: PaywallViewController,
-    isDebuggerLaunched: Bool,
     alertController: UIAlertController,
     completion: @escaping (SurveyPresentationResult) -> Void
   ) {
     alertController.dismiss(animated: true) {
       // Always complete without tracking if debugger launched.
-      if isDebuggerLaunched {
-        completion(.show)
-      } else {
-        Task {
-          let event = InternalSuperwallEvent.SurveyResponse(
-            survey: survey,
-            selectedOption: option,
-            customResponse: customResponse,
-            paywallInfo: paywallInfo
-          )
+      Task {
+        let event = InternalSuperwallEvent.SurveyResponse(
+          survey: survey,
+          selectedOption: option,
+          customResponse: customResponse,
+          paywallInfo: paywallInfo
+        )
+        await Superwall.shared.track(event)
 
-          let outcome = TrackingLogic.canTriggerPaywall(
-            event,
-            triggers: factory.makeTriggers(),
-            paywallViewController: paywallViewController
-          )
-          await Superwall.shared.track(event)
+        let outcome = TrackingLogic.canTriggerPaywall(
+          event,
+          triggers: factory.makeTriggers(),
+          paywallViewController: paywallViewController
+        )
 
-          // If we are going to implicitly show another paywall, we don't call the completion
-          // block as this will call didDismiss, which is going to be called
-          // implicitly anyway.
-          if outcome == .dontTriggerPaywall {
-            await MainActor.run {
-              completion(.show)
-            }
+        // If we are going to implicitly show another paywall, we don't call the completion
+        // block as this will call didDismiss, which is going to be called
+        // implicitly anyway.
+        if outcome == .dontTriggerPaywall {
+          await MainActor.run {
+            completion(.show)
           }
         }
       }