+++ Tests/SuperwallKitTests/Analytics/App Session/AppSessionManagerMock.swift
@@ -8,12 +8,15 @@
 import Foundation
 @testable import SuperwallKit
 
-class AppManagerDelegateMock: AppManagerDelegate, DeviceHelperFactory {
+class AppManagerDelegateMock: AppManagerDelegate, DeviceHelperFactory, UserAttributesEventFactory {
   func makeDeviceInfo() -> SuperwallKit.DeviceInfo {
     return .init(appInstalledAtString: "", locale: "")
   }
   func makeIsSandbox() -> Bool { return true}
   func makeSessionDeviceAttributes() async -> [String : Any] { [:] }
+  func makeUserAttributesEvent() -> InternalSuperwallEvent.Attributes { 
+    return InternalSuperwallEvent.Attributes(appInstalledAtString: "")
+  }
 
   func didUpdateAppSession(_ appSession: AppSession) async {}
 }
@@ -26,12 +29,14 @@ final class AppSessionManagerMock: AppSessionManager {
 
   init(
     appSession: AppSession,
+    identityManager: IdentityManager,
     configManager: ConfigManager,
     storage: Storage
   ) {
     internalAppSession = appSession
     super.init(
       configManager: configManager,
+      identityManager: identityManager,
       storage: storage,
       delegate: AppManagerDelegateMock()
     )