+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The changelog for `SuperwallKit`. Also see the [releases](https://github.com/sup
 - Tracks user attributes on session start.
 - Exposes `triggerSessionId` on the `PaywallInfo` object.
 - Makes `PaywallSkippedReason` conform to `CustomStringConvertible`.
+- Adds the Superwall SDK version and your app's version/build number to the debugger menu. Press the hamburger icon on the top left in the debugger to access it.
 
 ### Fixes